refactor: extract native modules
