+++ embrace-android-core/src/main/kotlin/io/embrace/android/embracesdk/internal/capture/envelope/resource/DeviceImpl.kt
@@ -11,6 +11,7 @@ import io.embrace.android.embracesdk.internal.isEmulator
 import io.embrace.android.embracesdk.internal.logging.EmbLogger
 import io.embrace.android.embracesdk.internal.logging.InternalErrorType
 import io.embrace.android.embracesdk.internal.prefs.PreferencesService
+import io.embrace.android.embracesdk.internal.utils.Provider
 import io.embrace.android.embracesdk.internal.worker.BackgroundWorker
 import java.io.File
 import java.util.Locale
@@ -20,7 +21,7 @@ public class DeviceImpl(
     private val preferencesService: PreferencesService,
     private val backgroundWorker: BackgroundWorker,
     override val systemInfo: SystemInfo,
-    cpuInfoDelegate: CpuInfoDelegate,
+    private val cpuInfoDelegate: Provider<CpuInfoDelegate?>,
     private val logger: EmbLogger
 ) : Device {
     override var isJailbroken: Boolean? = null
@@ -128,7 +129,7 @@ public class DeviceImpl(
     override val internalStorageTotalCapacity: Lazy<Long> =
         lazy { StatFs(Environment.getDataDirectory().path).totalBytes }
 
-    override val cpuName: String? = cpuInfoDelegate.getCpuName()
+    override val cpuName: String? by lazy { cpuInfoDelegate()?.getCpuName()}
 
-    override val eglInfo: String? = cpuInfoDelegate.getEgl()
+    override val eglInfo: String? by lazy { cpuInfoDelegate()?.getEgl() }
 }
