+++ embrace-android-sdk/src/main/java/io/embrace/android/embracesdk/internal/injection/NativeModuleImpl.kt
@@ -1,10 +1,13 @@
 package io.embrace.android.embracesdk.internal.injection
 
+import io.embrace.android.embracesdk.internal.SharedObjectLoader
 import io.embrace.android.embracesdk.internal.Systrace
 import io.embrace.android.embracesdk.internal.anr.ndk.EmbraceNativeThreadSamplerService
 import io.embrace.android.embracesdk.internal.anr.ndk.NativeAnrOtelMapper
 import io.embrace.android.embracesdk.internal.anr.ndk.NativeThreadSamplerInstaller
 import io.embrace.android.embracesdk.internal.anr.ndk.NativeThreadSamplerService
+import io.embrace.android.embracesdk.internal.capture.cpu.CpuInfoDelegate
+import io.embrace.android.embracesdk.internal.capture.cpu.EmbraceCpuInfoDelegate
 import io.embrace.android.embracesdk.internal.config.ConfigService
 import io.embrace.android.embracesdk.internal.ndk.EmbraceNdkService
 import io.embrace.android.embracesdk.internal.ndk.EmbraceNdkServiceRepository
@@ -39,7 +42,7 @@ internal class NativeModuleImpl(
                 essentialServiceModule.userService,
                 androidServicesModule.preferencesService,
                 essentialServiceModule.sessionPropertiesService,
-                essentialServiceModule.sharedObjectLoader,
+                sharedObjectLoader,
                 initModule.logger,
                 embraceNdkServiceRepository,
                 NdkDelegateImpl(),
@@ -60,7 +63,7 @@ internal class NativeModuleImpl(
                     logger = initModule.logger,
                     scheduledWorker = workerThreadModule.scheduledWorker(WorkerName.BACKGROUND_REGISTRATION),
                     deviceArchitecture = payloadSourceModule.deviceArchitecture,
-                    sharedObjectLoader = essentialServiceModule.sharedObjectLoader,
+                    sharedObjectLoader = sharedObjectLoader,
                 )
             } else {
                 null
@@ -76,7 +79,7 @@ internal class NativeModuleImpl(
         Systrace.traceSynchronous("native-thread-sampler-installer-init") {
             if (nativeThreadSamplingEnabled(configModule.configService)) {
                 NativeThreadSamplerInstaller(
-                    sharedObjectLoader = essentialServiceModule.sharedObjectLoader,
+                    sharedObjectLoader = sharedObjectLoader,
                     logger = initModule.logger
                 )
             } else {
@@ -99,6 +102,14 @@ internal class NativeModuleImpl(
                 logger = initModule.logger,
             )
         }
+    }
+
+    override val sharedObjectLoader: SharedObjectLoader by singleton {
+        SharedObjectLoader(initModule.logger)
+    }
+
+    override val cpuInfoDelegate: CpuInfoDelegate by singleton {
+        EmbraceCpuInfoDelegate(sharedObjectLoader, initModule.logger)
     }
 
     private fun nativeThreadSamplingEnabled(configService: ConfigService) = configService.autoDataCaptureBehavior.isNdkEnabled()
