+++ embrace-android-sdk/src/test/java/io/embrace/android/embracesdk/injection/EssentialServiceModuleImplTest.kt
@@ -50,11 +50,9 @@ internal class EssentialServiceModuleImplTest {
         assertNotNull(module.apiClient)
         assertNotNull(module.apiService)
         assertNotNull(module.activityLifecycleTracker)
-        assertNotNull(module.sharedObjectLoader)
         assertNotNull(module.sessionIdTracker)
         assertNotNull(module.sessionPropertiesService)
         assertTrue(module.userService is EmbraceUserService)
-        assertTrue(module.cpuInfoDelegate is EmbraceCpuInfoDelegate)
         assertTrue(module.networkConnectivityService is EmbraceNetworkConnectivityService)
         assertTrue(module.pendingApiCallsSender is EmbracePendingApiCallsSender)
         assertTrue(module.logWriter is LogWriterImpl)
